accounts/requirements.txt
Django==3.1.7
#psycopg2>=2.7,<3.0
mysqlclient==2.0.3
django-mysql>3.2.0
djangorestframework==3.12.2
django-rest-auth>=0.9.5
django-modeltranslation==0.16.2
requests==2.25.1
PyJWT==2.0.1
django-ckeditor>=6.0.0
django-oauth-toolkit==1.4.1
oauthlib>=3.1.0
django-cors-middleware>=1.5.0
phonenumbers==8.12.18
django-phonenumber-field==5.0.0
Pillow==8.2.0
git+https://git.kit-app.ru/integrationslibs/django-redsms-app.git


